+++ fpy2/interpret/__init__.py
@@ -2,7 +2,6 @@
 
 from .interpreter import Interpreter, get_default_interpreter, set_default_interpreter
 
-from .native import PythonInterpreter
 from .default import DefaultInterpreter
 
 set_default_interpreter(DefaultInterpreter())